Day350✅/requirements.txt
fastapi 
uvicorn 
langchain 
transformers
python-dotenv 
huggingface_hub